역할별로 파일 분리
